feat: address code review comments
